Write a cache sparse file.
Since the player seems to cache file size, a sparse file will present
the correct file size from the beginning on.

The problem is, now there is a random delay before xbmc actually starts
to play a file. I am not sure where it is coming from, possibly the
sparse file is confusing xbmc in some other way.